split up templates into files and gather in folders, move tests to folder
