+++ tests/templates/test_templates_embeddings.py
