+++ tests/templates/test_templates_layers.py
@@ -21,8 +21,8 @@ import pennylane as qml
 from pennylane import numpy as np
 from pennylane.templates.layers import (CVNeuralNetLayers,
                                         StronglyEntanglingLayers,
-                                        RandomLayers,
-                                        _random_layer)
+                                        RandomLayers)
+from pennylane.templates.layers.random import random_layer
 from pennylane import RX, RY, RZ, CZ, CNOT
 
 
@@ -61,7 +61,7 @@ class TestCVNeuralNet:
                          [ 0.26999146, 0.26256351, 0.14722687, 0.23137066]])
             ]
 
-    def test_cvneuralnet_uses_correct_weights(self, gaussian_device_4modes, weights):
+    def test_cvneuralnet_uses_correct_weights(self, weights):
         """Tests that the CVNeuralNetLayers template uses the weigh parameters correctly."""
 
         with qml.utils.OperationRecorder() as rec:
@@ -132,7 +132,6 @@ class TestStronglyEntangling:
         n_layers = 2
         num_wires = n_subsystems
 
-        dev = qml.device('default.qubit', wires=num_wires)
         weights = np.random.randn(n_layers, num_wires, 3)
 
         with qml.utils.OperationRecorder() as rec:
@@ -159,7 +158,6 @@ class TestStronglyEntangling:
     def test_strong_ent_layers_uses_correct_number_of_imprimitives(self, n_layers, n_subsystems):
         """Test that StronglyEntanglingLayers uses the correct number of imprimitives."""
         imprimitive = CZ
-        dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_layers, n_subsystems, 3)
 
         with qml.utils.OperationRecorder() as rec:
@@ -303,7 +301,6 @@ class TestRandomLayers:
         n_rots = 1
         n_wires = 2
         impr = CNOT
-        dev = qml.device('default.qubit', wires=n_wires)
         weights = np.random.randn(n_layers, n_rots)
 
         with qml.utils.OperationRecorder() as rec:
@@ -313,30 +310,28 @@ class TestRandomLayers:
         assert len(types) - types.count(impr) == n_layers
 
     def test_random_layer_ratio_imprimitive(self, ratio):
-        """Test that  _random_layer() has the right ratio of imprimitive gates."""
+        """Test that  random_layer() has the right ratio of imprimitive gates."""
         n_rots = 500
         n_wires = 2
         impr = CNOT
-        dev = qml.device('default.qubit', wires=n_wires)
         weights = np.random.randn(n_rots)
 
         with qml.utils.OperationRecorder() as rec:
-            _random_layer(weights=weights, wires=range(n_wires), ratio_imprim=ratio,
-                          imprimitive=CNOT, rotations=[RX, RY, RZ], seed=42)
+            random_layer(weights=weights, wires=range(n_wires), ratio_imprim=ratio,
+                         imprimitive=CNOT, rotations=[RX, RY, RZ], seed=42)
 
         types = [type(q) for q in rec.queue]
         ratio_impr = types.count(impr) / len(types)
         assert np.isclose(ratio_impr, ratio, atol=0.05)
 
     def test_random_layer_gate_types(self, n_subsystems, impr, rots):
-        """Test that  _random_layer() uses the correct types of gates."""
+        """Test that  random_layer() uses the correct types of gates."""
         n_rots = 20
-        dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_rots)
 
         with qml.utils.OperationRecorder() as rec:
-            _random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
-                          imprimitive=impr, rotations=rots, seed=42)
+            random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
+                         imprimitive=impr, rotations=rots, seed=42)
 
         types = [type(q) for q in rec.queue]
         unique = set(types)
@@ -344,27 +339,25 @@ class TestRandomLayers:
         assert unique == gates
 
     def test_random_layer_numgates(self, n_subsystems):
-        """Test that _random_layer() uses the correct number of gates."""
+        """Test that random_layer() uses the correct number of gates."""
         n_rots = 5
-        dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_rots)
 
         with qml.utils.OperationRecorder() as rec:
-            _random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
-                          imprimitive=qml.CNOT, rotations=[RX, RY, RZ], seed=42)
+            random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
+                         imprimitive=qml.CNOT, rotations=[RX, RY, RZ], seed=42)
 
         types = [type(q) for q in rec.queue]
         assert len(types) - types.count(qml.CNOT) == n_rots
 
     def test_random_layer_randomwires(self, n_subsystems):
-        """Test that  _random_layer() picks random wires."""
+        """Test that  random_layer() picks random wires."""
         n_rots = 500
-        dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_rots)
 
         with qml.utils.OperationRecorder() as rec:
-            _random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
-                          imprimitive=qml.CNOT, rotations=[RX, RY, RZ], seed=42)
+            random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
+                         imprimitive=qml.CNOT, rotations=[RX, RY, RZ], seed=42)
 
         wires = [q._wires for q in rec.queue]
         wires_flat = [item for w in wires for item in w]
@@ -372,15 +365,14 @@ class TestRandomLayers:
         assert np.isclose(mean_wire, (n_subsystems - 1) / 2, atol=0.05)
 
     def test_random_layer_weights(self, n_subsystems, tol):
-        """Test that _random_layer() uses the correct weights."""
+        """Test that random_layer() uses the correct weights."""
         np.random.seed(12)
         n_rots = 5
-        dev = qml.device('default.qubit', wires=n_subsystems)
         weights = np.random.randn(n_rots)
 
         with qml.utils.OperationRecorder() as rec:
-            _random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
-                          imprimitive=qml.CNOT, rotations=[RX, RY, RZ], seed=4)
+            random_layer(weights=weights, wires=range(n_subsystems), ratio_imprim=0.3,
+                         imprimitive=qml.CNOT, rotations=[RX, RY, RZ], seed=4)
 
         params = [q.parameters for q in rec.queue]
         params_flat = [item for p in params for item in p]
